requirements.txt
requests>=2.32,<3
python-dotenv>=1.0,<2
tzdata>=2025.2
pytest>=8.0,<9
